Added backend authentication and global exception handling

+++ main/java/com/gotham/batman/controllers/CandidateController.java
@@ -1,6 +1,7 @@
 package com.gotham.batman.controllers;
 
 import java.util.List;
+import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +11,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
@@ -26,55 +28,82 @@ public class CandidateController {
 	private final static Logger LOGGER =  Logger.getLogger(Logger.GLOBAL_LOGGER_NAME);
 	@RequestMapping(value="/addcandidate", method=RequestMethod.POST)
 	
-	public String addCandidate(@RequestBody Candidate candidate)
+	public String addCandidate(@RequestBody Candidate candidate,@RequestHeader("Authorization") String token)
 	{
-		//System.out.println(candidate.toString());
-		//candidateService.addCandidate(new Candidate("harish","sundar","harish@gmail.com","cadvd","sdsd","vvewv","wewre"));
+		if(candidateService.checkUser(token)=="Login sucessfull and User is Authenticated") {
+		LOGGER.log(Level.INFO, "Candidate Added	"); 
 		candidateService.addCandidate(candidate);
 		return "completed";
+		}
+		return null;
 	}
 	
 	@DeleteMapping("/deletecandidate/{Id}")
-	public List<Candidate> deleteCandidate(@PathVariable Integer Id)
+	public List<Candidate> deleteCandidate(@PathVariable Integer Id,@RequestHeader("Authorization") String token)
 	{
+		if(candidateService.checkUser(token)=="Login sucessfull and User is Authenticated") {
+		LOGGER.log(Level.INFO, "Candidate Deleted	");
 		candidateService.deleteCandidate(Id);
 		return candidateService.getAllCandidates();
+		}
+		return null;
 	}
 	
 	@GetMapping("/getcandidatebyid/{Id}")
-	public List<Candidate> getCandidateById(@PathVariable Integer Id)
+	public List<Candidate> getCandidateById(@PathVariable Integer Id,@RequestHeader("Authorization") String token)
 	{
+		if(candidateService.checkUser(token)=="Login sucessfull and User is Authenticated") {
 		return candidateService.getCandidateById(Id);
+		}
+		return null;
 	}
 	@GetMapping("/getuserbyid/{Id}")
-	public Candidate getUserById(@PathVariable Integer Id)
+	public Candidate getUserById(@PathVariable Integer Id,@RequestHeader("Authorization") String token)
 	{
+		if(candidateService.checkUser(token)=="Login sucessfull and User is Authenticated") {
 		return candidateService.getUserById(Id);
+		}
+		return null;
 	}
 	
 	@GetMapping("/getcandidatebylocation/{location}")
-	public List<Candidate> getCandidateByLoc(@PathVariable String location)
+	public List<Candidate> getCandidateByLoc(@PathVariable String location,@RequestHeader("Authorization") String token)
 	{
+		if(candidateService.checkUser(token)=="Login sucessfull and User is Authenticated") {
 		return candidateService.getCandidateByLocation(location);
+		}
+		return null;
 	}
 	
 	@GetMapping("/getallcandidates")
-	public List<Candidate> getAllCandidates()
+	public List<Candidate> getAllCandidates(@RequestHeader("Authorization") String token)
 	{
+		if(candidateService.checkUser(token)=="Login sucessfull and User is Authenticated") {
 		return candidateService.getAllCandidates();
+		}
+		return null;
 	}
 	
 	@GetMapping("/getlocationcount")
-	public List<LocationCount> getLocationCount()
-	{
+	public List<LocationCount> getLocationCount(@RequestHeader("Authorization") String token)
+	{if(candidateService.checkUser(token)=="Login sucessfull and User is Authenticated") {
 		return candidateService.getLocation();
 	}
+	return null;
+	}
 	@PostMapping("/updatecandidate/{id}")
-	public List<Candidate> updateCandidate(@RequestBody Candidate candidate,@PathVariable Integer id)
+	public List<Candidate> updateCandidate(@RequestBody Candidate candidate,@PathVariable Integer id,@RequestHeader("Authorization") String token)
 	{
-		System.out.println(id);
+		if(candidateService.checkUser(token)=="Login sucessfull and User is Authenticated") {
+		LOGGER.log(Level.INFO, "Candidate Updated	");
 		candidateService.updateCandidate(new Candidate(id,candidate.getFirstName(),candidate.getLastName(),candidate.getEmail(),candidate.getLocation(),candidate.getFeedback(),candidate.getJobDescription(),candidate.getContactNumber()));
 		return candidateService.getAllCandidates();
+		}
+		return null;
 	}
-	
+	@GetMapping("/users")
+	String checkUser(@RequestHeader("Authorization") String token)
+	{
+			return candidateService.checkUser(token);
+	}
 }
